requirements.txt
Django==4.2
python-dotenv==1.0.0
python-telegram-bot==13.2.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
qrcode==7.4.2